pyproject.toml
[tool.poetry]
name = "estante-digital"
version = "0.1.0"
description = ""
authors = ["Pedro <pedro5@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tzdata = "^2024.1"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
taskipy = "^1.12.2"
ruff = "^0.4.5"
httpx = "^0.27.0"
factory-boy = "^3.3.0"
freezegun = "^1.5.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev estante_digital/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=estante_digital -vv'
post_test = 'coverage html'
migrate = 'alembic revision --autogenerate'
upgrade = 'alembic upgrade head'
downgrade = 'alembic downgrade'

[tool.coverage.report]
exclude_also = ['def get_session()']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
